requirements.txt
streamlit==1.28.1
google-generativeai==0.3.0
PyPDF2==3.0.1
pandas==2.0.3
python-dotenv==1.0.0
